+++ tests/testthat/test-colwise-select.R
@@ -145,11 +145,11 @@ test_that("select_all does not change the order of columns (#3351)", {
 
 test_that("mutate_all does not change the order of columns (#3351)", {
   tbl <- group_by(tibble(x = 1:4, y = 1:4), y)
-  expect_identical(names(mutate_all(tbl, identity)), names(tbl))
+  expect_warning(expect_identical(names(mutate_all(tbl, identity)), names(tbl)), "grouped tibble")
 
   tbl <- group_by(tibble(x = 1:4, y = 1:4), x)
-  expect_identical(names(mutate_all(tbl, identity)), names(tbl))
+  expect_warning(expect_identical(names(mutate_all(tbl, identity)), names(tbl)), "grouped tibble")
 
   tbl <- group_by(tibble(x = 1:4, y = 1:4, z = 1:4), y)
-  expect_identical(names(mutate_all(tbl, identity)), names(tbl))
+  expect_warning(expect_identical(names(mutate_all(tbl, identity)), names(tbl)), "grouped tibble")
 })